Move debouncer to its own module
